Fixed types issues, types file now include TopBar props
