+++ src/components/modal/types.ts
@@ -1,7 +1,21 @@
-import {ModalProps as RNModalProps, GestureResponderEvent, KeyboardAvoidingViewProps} from 'react-native';
-import {ModalTopBarProps} from './TopBar';
+import {
+  ModalProps as RNModalProps,
+  GestureResponderEvent,
+  KeyboardAvoidingViewProps,
+  StyleProp,
+  TextStyle,
+  ImageSourcePropType
+} from 'react-native';
+import {ButtonProps} from '../../components/button';
+import {ViewProps} from '../../components/view';
 
-export {ModalTopBarProps};
+export type topBarButtonProp = {
+  onPress?: (props: any) => void;
+  label?: string;
+  icon?: ImageSourcePropType;
+  accessibilityLabel?: string;
+  buttonProps?: Omit<ButtonProps, 'onPress'>;
+};
 
 export interface ModalProps extends RNModalProps {
   /**
@@ -45,4 +59,67 @@ export interface ModalProps extends RNModalProps {
    * The modal's container identifier for web implementation, the container id will be the parent element/div for the modal's to render in.
    */
   portalContainerId?: string;
+}
+
+export interface ModalTopBarProps {
+  /**
+   * title to display in the center of the top bar
+   */
+  title?: string;
+  /**
+   * title custom style
+   */
+  titleStyle?: StyleProp<TextStyle>;
+  /**
+   * done action props (Button props)
+   */
+  doneButtonProps?: Omit<ButtonProps, 'onPress'>;
+  /**
+   * done action label
+   */
+  doneLabel?: string;
+  /**
+   * done action icon
+   */
+  doneIcon?: ImageSourcePropType;
+  /**
+   * done action callback
+   */
+  onDone?: (props?: any) => void;
+  /**
+   * cancel action props (Button props)
+   */
+  cancelButtonProps?: Omit<ButtonProps, 'onPress'>;
+  /**
+   * cancel action label
+   */
+  cancelLabel?: string;
+  /**
+   * cancel action icon
+   */
+  cancelIcon?: ImageSourcePropType;
+  /**
+   * cancel action callback
+   */
+  onCancel?: (props?: any) => void;
+  /**
+   * buttons to render on the right side of the top bar
+   */
+  rightButtons?: topBarButtonProp | topBarButtonProp[];
+  /**
+   * buttons to render on the left side of the top bar
+   */
+  leftButtons?: topBarButtonProp | topBarButtonProp[];
+  /**
+   * whether to include status bar or not (height claculations)
+   */
+  includeStatusBar?: boolean;
+  /**
+   * style for the TopBar container
+   */
+  containerStyle?: ViewProps['style'];
+  /**
+   * Whether or not to handle SafeArea
+   */
+  useSafeArea?: boolean;
 }
